auth0-be/python-auth0-hello-world/requirements.txt
cumulio==0.2.3
Flask==2.1.2
PyJWT==2.5.0
python-dotenv==0.21.0
Flask-Cors==3.0.10